refactor: "uint64_t" and "int64_t" instead of "long long"

+++ kernel/class.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include "class.h"
 
-class_t *constructor_class(long long index_class, char *class_name, long long number_field){
+class_t *constructor_class(int index_class, char *class_name, int number_field){
     class_t *class = malloc(sizeof(class_t));
 
     class->index_class = index_class;
@@ -22,7 +22,7 @@ void destructor_class(class_t *class) {
 
 object_t *constructor_object(class_t *class) {
     object_t *object = malloc(sizeof(object_t));
-    object->fields = malloc(sizeof(long long) * class->number_fields);
+    object->fields = malloc(sizeof(int64_t) * class->number_fields);
     object->is_field_object= calloc(1, sizeof(bool) * class->number_fields);
     object->class = class;
     return object;
